Create 643-fixed.c
This is a reverse shell buffer overflow written for SLMail version 5.5 in C. It works and it took me a lot work to modify since I don't normally code in C. Seriously, this is malware. Don't use it unless you know what you're doing and you have permission to pen test the systems you are trying to break!