Create 646-fixed.c
This is a reverse shell buffer overflow for SLMail Version 5.5. It was originally written in C for Windows. I modified it to compile in Linux and be executed using Wine. Thank you MingW. Not something you should run against your grandmother's SLMail server, unless you are a real jerk. Malware. Use at your own discretion and only against systems for which you are the admin.